requirements.txt
base58==2.1.1
bitcoinlib==0.6.12
certifi==2023.11.17
charset-normalizer==3.3.2
ecdsa==0.18.0
fastecdsa==2.3.0
greenlet==3.0.1
idna==3.4
numpy==1.26.2
pycryptodome==3.19.0
requests==2.31.0
ripemd-hash==1.0.1
six==1.16.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
